rename package 'commands' to 'command'
