+++ core/src/main/java/pl/mrstudios/deathrun/plugin/Bootstrap.java
@@ -25,9 +25,9 @@ import pl.mrstudios.deathrun.arena.trap.TrapRegistry;
 import pl.mrstudios.deathrun.arena.trap.impl.TrapAppearingBlocks;
 import pl.mrstudios.deathrun.arena.trap.impl.TrapDisappearingBlocks;
 import pl.mrstudios.deathrun.arena.trap.impl.TrapTNT;
-import pl.mrstudios.deathrun.commands.CommandDeathRun;
-import pl.mrstudios.deathrun.commands.handler.InvalidCommandUsageHandler;
-import pl.mrstudios.deathrun.commands.handler.NoCommandPermissionsHandler;
+import pl.mrstudios.deathrun.command.CommandDeathRun;
+import pl.mrstudios.deathrun.command.handler.InvalidCommandUsageHandler;
+import pl.mrstudios.deathrun.command.handler.NoCommandPermissionsHandler;
 import pl.mrstudios.deathrun.config.Configuration;
 import pl.mrstudios.deathrun.config.ConfigurationFactory;
 import pl.mrstudios.deathrun.config.impl.LanguageConfiguration;
